requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
gspread
python-dotenv
requests
googlesearch-python
